GTM Engine/backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0
aiosqlite==0.20.0
alembic==1.13.1

# Data validation
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# Auth
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
python-multipart==0.0.9

# Background tasks
celery==5.4.0
redis==5.0.4

# HTTP client
httpx==0.27.0

# AI
anthropic==0.28.0

# Integrations
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.130.0
hubspot-api-client==8.2.1
slack-sdk==3.27.2

# Utilities
python-dotenv==1.0.1
structlog==24.1.0
tenacity==8.3.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
pytest-cov==5.0.0

# Dev tools
black==24.4.2
isort==5.13.2
